#776 Remove implementation of ModelBase.Equals and ModelBase.GetHashCode to improve out-of-the-box performance of the ModelBase. To implement comparison, one has to override it manually or derive from ComparableModelBase
