+++ src/Catel.Core/Catel.Core.Shared/Data/ModelBase.cs
@@ -67,22 +67,6 @@ namespace Catel.Data
 #endif
         private bool _leanAndMeanModel;
 
-        /// <summary>
-        /// Backing field for the <see cref="EqualityComparer"/> property. Because it has custom logic, it needs a backing field.
-        /// </summary>
-#if NET
-        [field: NonSerialized]
-#endif
-        private IModelEqualityComparer _equalityComparer;
-
-        /// <summary>
-        /// Backing field for the <see cref="GetHashCode"/> method so it only has to be calculated once to gain the best performance possible.
-        /// </summary>
-#if NET
-        [field: NonSerialized]
-#endif
-        private int? _hashCode;
-
 #if NET
         [field: NonSerialized]
 #endif
@@ -134,59 +118,7 @@ namespace Catel.Data
 #endif
         #endregion
 
-        #region Operators
-        /// <summary>
-        /// Implements the operator ==.
-        /// </summary>
-        /// <param name="firstObject">The first object.</param>
-        /// <param name="secondObject">The second object.</param>
-        /// <returns>The result of the operator.</returns>
-        public static bool operator ==(ModelBase firstObject, ModelBase secondObject)
-        {
-            return Equals(firstObject, secondObject);
-        }
-
-        /// <summary>
-        /// Implements the operator !=.
-        /// </summary>
-        /// <param name="firstObject">The first object.</param>
-        /// <param name="secondObject">The second object.</param>
-        /// <returns>The result of the operator.</returns>
-        public static bool operator !=(ModelBase firstObject, ModelBase secondObject)
-        {
-            return !(firstObject == secondObject);
-        }
-
-        #endregion
-
         #region Properties
-        /// <summary>
-        /// Gets or sets the equality comparer used to compare model bases with each other.
-        /// </summary>
-        /// <value>The equality comparer.</value>
-#if NET
-        [Browsable(false)]
-#endif
-        [XmlIgnore]
-        protected IModelEqualityComparer EqualityComparer
-        {
-            get
-            {
-                if (_equalityComparer == null)
-                {
-                    var dependencyResolver = this.GetDependencyResolver();
-
-                    _equalityComparer = dependencyResolver.Resolve<IModelEqualityComparer>();
-                }
-
-                return _equalityComparer;
-            }
-            set
-            {
-                _equalityComparer = value;
-            }
-        }
-
         /// <summary>
         /// Gets or sets a value indicating whether all models should behave as a lean and mean model.
         /// <para />
@@ -378,54 +310,7 @@ namespace Catel.Data
             InitializeCustomProperties();
         }
 
-        /// <summary>
-        /// Determines whether the specified <see cref="System.Object"/> is equal to this instance.
-        /// </summary>
-        /// <param name="obj">The <see cref="System.Object"/> to compare with this instance.</param>
-        /// <returns>
-        /// <c>true</c> if the specified <see cref="System.Object"/> is equal to this instance; otherwise, <c>false</c>.
-        /// </returns>
-        /// <exception cref="T:System.NullReferenceException">
-        /// The <paramref name="obj"/> parameter is null.
-        /// </exception>
-        public override bool Equals(object obj)
-        {
-            // Note: at first we only implemented the EqualityComparer, but the IEqualityComparer of Microsoft
-            // throws an exception when the 2 types are not the same. Although MS does recommend not to throw exceptions,
-            // they do it themselves. Check for null and check the types before feeding it to the equality comparer.
-
-            if (obj == null)
-            {
-                return false;
-            }
-
-            if (GetType() != obj.GetType())
-            {
-                return false;
-            }
-
-            var equalityComparer = EqualityComparer;
-            return equalityComparer.Equals(this, obj);
-        }
-
         // ReSharper disable RedundantOverridenMember
-
-        /// <summary>
-        /// Returns a hash code for this instance.
-        /// </summary>
-        /// <returns>
-        /// A hash code for this instance, suitable for use in hashing algorithms and data structures like a hash table. 
-        /// </returns>
-        public override int GetHashCode()
-        {
-            if (!_hashCode.HasValue)
-            {
-                var equalityComparer = EqualityComparer;
-                _hashCode = equalityComparer.GetHashCode(this);
-            }
-
-            return _hashCode.Value;
-        }
 
         /// <summary>
         /// Returns a <see cref="System.String"/> that represents this instance.
