+++ src/Catel.Test/Catel.Test.Shared/Data/TestClasses/IniEntry.cs
@@ -12,7 +12,7 @@ namespace Catel.Test.Data
 #if NET
     [Serializable]
 #endif
-    public class IniEntry : SavableModelBase<IniEntry>
+    public class IniEntry : ComparableModelBase
     {
         #region Fields
         #endregion
